+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## NEXT
 
 * Minor tweaks to Pushover notification events
+* [FIX] Plugin doesn't start properly when controller in Settings mode 
 
 ## 3.2.0 (2020-11-01)